Fixed numpy compatibility bug

+++ src/analysis/dist_helper.py
@@ -12,12 +12,12 @@ from scipy.linalg import toeplitz
 
 def emd(x, y, distance_scaling=1.0):
     support_size = max(len(x), len(y))
-    d_mat = toeplitz(range(support_size)).astype(np.float)
+    d_mat = toeplitz(range(support_size)).astype(float)
     distance_mat = d_mat / distance_scaling
 
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     if len(x) < len(y):
         x = np.hstack((x, [0.0] * (support_size - len(x))))
     elif len(y) < len(x):
@@ -40,12 +40,12 @@ def emd(x, y, sigma=1.0, distance_scaling=1.0):
             sigma: standard deviation
     '''
     support_size = max(len(x), len(y))
-    d_mat = toeplitz(range(support_size)).astype(np.float)
+    d_mat = toeplitz(range(support_size)).astype(float)
     distance_mat = d_mat / distance_scaling
 
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     if len(x) < len(y):
         x = np.hstack((x, [0.0] * (support_size - len(x))))
     elif len(y) < len(x):
@@ -61,12 +61,12 @@ def gaussian_emd(x, y, sigma=1.0, distance_scaling=1.0):
             sigma: standard deviation
     '''
     support_size = max(len(x), len(y))
-    d_mat = toeplitz(range(support_size)).astype(np.float)
+    d_mat = toeplitz(range(support_size)).astype(float)
     distance_mat = d_mat / distance_scaling
 
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     if len(x) < len(y):
         x = np.hstack((x, [0.0] * (support_size - len(x))))
     elif len(y) < len(x):
@@ -79,8 +79,8 @@ def gaussian_emd(x, y, sigma=1.0, distance_scaling=1.0):
 def gaussian(x, y, sigma=1.0):
     support_size = max(len(x), len(y))
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     if len(x) < len(y):
         x = np.hstack((x, [0.0] * (support_size - len(x))))
     elif len(y) < len(x):
@@ -93,8 +93,8 @@ def gaussian(x, y, sigma=1.0):
 def gaussian_tv(x, y, sigma=1.0):  
     support_size = max(len(x), len(y))
     # convert histogram values x and y to float, and make them equal len
-    x = x.astype(np.float)
-    y = y.astype(np.float)
+    x = x.astype(float)
+    y = y.astype(float)
     if len(x) < len(y):
         x = np.hstack((x, [0.0] * (support_size - len(x))))
     elif len(y) < len(x):
